ia-backend/requirements.txt
python-dotenv
Flask
flask-cors
openai==0.28
psycopg2-binary